requirements.txt
python=3.8.13
streamlit == 1.12.2
python-docx==0.8.11